Remove DiffViewer, rename PulsarChart → RidgelineChart, bump 0.5.0
DiffViewer ships a 9.7 MB bundle from @pierre/diffs's 298 TextMate
grammars; removing it drops the static dir from ~12.5 MB to 2.9 MB
(#217). PulsarChart is renamed to RidgelineChart, the conventional
name for stacked-waveform "Joy Division" plots (#87). Both are
breaking changes, so version bumps to 0.5.0.
