+++ README.md
@@ -46,7 +46,7 @@ uv pip install wigglystuff
 </tr>
 <tr>
 <td align="center"><b>ProgressBar</b><br><a href="https://molab.marimo.io/github/koaning/wigglystuff/blob/main/demos/progressbar.py/wasm"><img src="./docs/assets/gallery/progressbar.webp" width="330"></a><br><a href="https://molab.marimo.io/github/koaning/wigglystuff/blob/main/demos/progressbar.py/wasm">molab</a> · <a href="https://koaning.github.io/wigglystuff/reference/progress-bar/">API</a> · <a href="https://koaning.github.io/wigglystuff/reference/progress-bar.md">MD</a></td>
-<td align="center"><b>PulsarChart</b><br><a href="https://molab.marimo.io/github/koaning/wigglystuff/blob/main/demos/pulsarchart.py/wasm"><img src="./docs/assets/gallery/pulsarchart.webp" width="330"></a><br><a href="https://molab.marimo.io/github/koaning/wigglystuff/blob/main/demos/pulsarchart.py/wasm">molab</a> · <a href="https://koaning.github.io/wigglystuff/reference/pulsar-chart/">API</a> · <a href="https://koaning.github.io/wigglystuff/reference/pulsar-chart.md">MD</a></td>
+<td align="center"><b>RidgelineChart</b><br><a href="https://molab.marimo.io/github/koaning/wigglystuff/blob/main/demos/ridgelinechart.py/wasm"><img src="./docs/assets/gallery/ridgelinechart.webp" width="330"></a><br><a href="https://molab.marimo.io/github/koaning/wigglystuff/blob/main/demos/ridgelinechart.py/wasm">molab</a> · <a href="https://koaning.github.io/wigglystuff/reference/ridgeline-chart/">API</a> · <a href="https://koaning.github.io/wigglystuff/reference/ridgeline-chart.md">MD</a></td>
 <td align="center"><b>TextCompare</b><br><a href="https://molab.marimo.io/github/koaning/wigglystuff/blob/main/demos/textcompare.py/wasm"><img src="./docs/assets/gallery/textcompare.webp" width="330"></a><br><a href="https://molab.marimo.io/github/koaning/wigglystuff/blob/main/demos/textcompare.py/wasm">molab</a> · <a href="https://koaning.github.io/wigglystuff/reference/text-compare/">API</a> · <a href="https://koaning.github.io/wigglystuff/reference/text-compare.md">MD</a></td>
 <td align="center"><b>EnvConfig</b><br><a href="https://molab.marimo.io/github/koaning/wigglystuff/blob/main/demos/envconfig.py/wasm"><img src="./docs/assets/gallery/envconfig.webp" width="330"></a><br><a href="https://molab.marimo.io/github/koaning/wigglystuff/blob/main/demos/envconfig.py/wasm">molab</a> · <a href="https://koaning.github.io/wigglystuff/reference/env-config/">API</a> · <a href="https://koaning.github.io/wigglystuff/reference/env-config.md">MD</a></td>
 </tr>
@@ -58,7 +58,6 @@ uv pip install wigglystuff
 </tr>
 <tr>
 <td align="center"><b>ScatterWidget</b><br><a href="https://molab.marimo.io/github/koaning/wigglystuff/blob/main/demos/scatterwidget.py/wasm"><img src="./docs/assets/gallery/scatterwidget.webp" width="330"></a><br><a href="https://molab.marimo.io/github/koaning/wigglystuff/blob/main/demos/scatterwidget.py/wasm">molab</a> · <a href="https://koaning.github.io/wigglystuff/reference/scatter-widget/">API</a> · <a href="https://koaning.github.io/wigglystuff/reference/scatter-widget.md">MD</a></td>
-<td align="center"><b>DiffViewer</b><br><a href="https://molab.marimo.io/github/koaning/wigglystuff/blob/main/demos/diffviewer.py/wasm"><img src="./docs/assets/gallery/diffviewer.webp" width="330"></a><br><a href="https://molab.marimo.io/github/koaning/wigglystuff/blob/main/demos/diffviewer.py/wasm">molab</a> · <a href="https://koaning.github.io/wigglystuff/reference/diff-viewer/">API</a> · <a href="https://koaning.github.io/wigglystuff/reference/diff-viewer.md">MD</a></td>
 <td align="center"><b>SplineDraw</b><br><a href="https://molab.marimo.io/github/koaning/wigglystuff/blob/main/demos/splinedraw.py/wasm"><img src="./docs/assets/gallery/splinedraw.webp" width="330"></a><br><a href="https://molab.marimo.io/github/koaning/wigglystuff/blob/main/demos/splinedraw.py/wasm">molab</a> · <a href="https://koaning.github.io/wigglystuff/reference/spline-draw/">API</a> · <a href="https://koaning.github.io/wigglystuff/reference/spline-draw.md">MD</a></td>
 <td align="center"><b>ApiDoc</b><br><a href="https://molab.marimo.io/github/koaning/wigglystuff/blob/main/demos/apidoc.py/wasm"><img src="./docs/assets/gallery/apidoc.webp" width="330"></a><br><a href="https://molab.marimo.io/github/koaning/wigglystuff/blob/main/demos/apidoc.py/wasm">molab</a> · <a href="https://koaning.github.io/wigglystuff/reference/api-doc/">API</a> · <a href="https://koaning.github.io/wigglystuff/reference/api-doc.md">MD</a></td>
 </tr>
