+++ docs/llms.txt
@@ -29,11 +29,10 @@ Install via `pip install wigglystuff` or `uv pip install wigglystuff`.
 - [ImageRefreshWidget](https://koaning.github.io/wigglystuff/reference/image-refresh.md): Auto-refreshing images
 - [HTMLRefreshWidget](https://koaning.github.io/wigglystuff/reference/html-refresh.md): Auto-refreshing HTML content
 - [ProgressBar](https://koaning.github.io/wigglystuff/reference/progress-bar.md): Progress indicator widget
-- [PulsarChart](https://koaning.github.io/wigglystuff/reference/pulsar-chart.md): Stacked waveform visualization (Joy Division style)
+- [RidgelineChart](https://koaning.github.io/wigglystuff/reference/ridgeline-chart.md): Stacked waveform visualization (Joy Division style)
 - [TextCompare](https://koaning.github.io/wigglystuff/reference/text-compare.md): Side-by-side text comparison with match highlighting
 - [EnvConfig](https://koaning.github.io/wigglystuff/reference/env-config.md): Environment variable configuration with validation
 - [ScatterWidget](https://koaning.github.io/wigglystuff/reference/scatter-widget.md): Interactive scatter drawing for painting labeled 2D datasets (re-exported from drawdata)
-- [DiffViewer](https://koaning.github.io/wigglystuff/reference/diff-viewer.md): Rich file diff viewer with split/unified modes powered by @pierre/diffs
 - [SplineDraw](https://koaning.github.io/wigglystuff/reference/spline-draw.md): Draw scatter points with Python-computed spline curve fitting
 - [ApiDoc](https://koaning.github.io/wigglystuff/reference/api-doc.md): Renders API docs for Python classes and functions directly in notebooks
 - [PlaySlider](https://koaning.github.io/wigglystuff/reference/play-slider.md): Slider with play/pause button for auto-advancing values at a configurable pace
