+++ package.json
@@ -11,7 +11,6 @@
   },
   "dependencies": {
     "@anywidget/react": "^0.1.0",
-    "@pierre/diffs": "^1.0.11",
     "d3-array": "^3.2.4",
     "d3-axis": "^3.0.0",
     "d3-color": "^3.1.0",
