--- wigglystuff/static/pulsar-chart.js
@@ -1,1 +1,0 @@
-var $t="http://www.w3.org/1999/xhtml",ee={svg:"http://www.w3.org/2000/svg",xhtml:$t,xlink:"http://www.w3.org/1999/xlink",xml:"http://www.w3.org/XML/1998/namespace",xmlns:"http://www.w3.org/2000/xmlns/"};function St(t){var e=t+="",r=e.indexOf(":");return r>=0&&(e=t.slice(0,r))!=="xmlns"&&(t=t.slice(r+1)),ee.hasOwnProperty(e)?{space:ee[e],local:t}:t}function dn(t){return function(){var e=this.ownerDocument,r=this.namespaceURI;return r===$t&&e.documentElement.namespaceURI===$t?e.createElement(t):e.createElementNS(r,t)}}function mn(t){return function(){return this.ownerDocument.createElementNS(t.space,t.local)}}function Nt(t){var e=St(t);return(e.local?mn:dn)(e)}function hn(){}function Rt(t){return t==null?hn:function(){return this.querySelector(t)}}function Be(t){typeof t!="function"&&(t=Rt(t));for(var e=this._groups,r=e.length,n=new Array(r),o=0;o<r;++o)for(var i=e[o],a=i.length,f=n[o]=new Array(a),l,s,p=0;p<a;++p)(l=i[p])&&(s=t.call(l,l.__data__,p,i))&&("__data__"in l&&(s.__data__=l.__data__),f[p]=s);return new $(n,this._parents)}function re(t){return t==null?[]:Array.isArray(t)?t:Array.from(t)}function xn(){return[]}function Fe(t){return t==null?xn:function(){return this.querySelectorAll(t)}}function gn(t){return function(){return re(t.apply(this,arguments))}}function De(t){typeof t=="function"?t=gn(t):t=Fe(t);for(var e=this._groups,r=e.length,n=[],o=[],i=0;i<r;++i)for(var a=e[i],f=a.length,l,s=0;s<f;++s)(l=a[s])&&(n.push(t.call(l,l.__data__,s,a)),o.push(l));return new $(n,o)}function Te(t){return function(){return this.matches(t)}}function Ct(t){return function(e){return e.matches(t)}}var yn=Array.prototype.find;function bn(t){return function(){return yn.call(this.children,t)}}function wn(){return this.firstElementChild}function ze(t){return this.select(t==null?wn:bn(typeof t=="function"?t:Ct(t)))}var _n=Array.prototype.filter;function vn(){return Array.from(this.children)}function kn(t){return function(){return _n.call(this.children,t)}}function Ve(t){return this.selectAll(t==null?vn:kn(typeof t=="function"?t:Ct(t)))}function Xe(t){typeof t!="function"&&(t=Te(t));for(var e=this._groups,r=e.length,n=new Array(r),o=0;o<r;++o)for(var i=e[o],a=i.length,f=n[o]=[],l,s=0;s<a;++s)(l=i[s])&&t.call(l,l.__data__,s,i)&&f.push(l);return new $(n,this._parents)}function Pt(t){return new Array(t.length)}function Ye(){return new $(this._enter||this._groups.map(Pt),this._parents)}function pt(t,e){this.ownerDocument=t.ownerDocument,this.namespaceURI=t.namespaceURI,this._next=null,this._parent=t,this.__data__=e}pt.prototype={constructor:pt,appendChild:function(t){return this._parent.insertBefore(t,this._next)},insertBefore:function(t,e){return this._parent.insertBefore(t,e)},querySelector:function(t){return this._parent.querySelector(t)},querySelectorAll:function(t){return this._parent.querySelectorAll(t)}};function Ge(t){return function(){return t}}function Mn(t,e,r,n,o,i){for(var a=0,f,l=e.length,s=i.length;a<s;++a)(f=e[a])?(f.__data__=i[a],n[a]=f):r[a]=new pt(t,i[a]);for(;a<l;++a)(f=e[a])&&(o[a]=f)}function An(t,e,r,n,o,i,a){var f,l,s=new Map,p=e.length,u=i.length,c=new Array(p),d;for(f=0;f<p;++f)(l=e[f])&&(c[f]=d=a.call(l,l.__data__,f,e)+"",s.has(d)?o[f]=l:s.set(d,l));for(f=0;f<u;++f)d=a.call(t,i[f],f,i)+"",(l=s.get(d))?(n[f]=l,l.__data__=i[f],s.delete(d)):r[f]=new pt(t,i[f]);for(f=0;f<p;++f)(l=e[f])&&s.get(c[f])===l&&(o[f]=l)}function $n(t){return t.__data__}function Ue(t,e){if(!arguments.length)return Array.from(this,$n);var r=e?An:Mn,n=this._parents,o=this._groups;typeof t!="function"&&(t=Ge(t));for(var i=o.length,a=new Array(i),f=new Array(i),l=new Array(i),s=0;s<i;++s){var p=n[s],u=o[s],c=u.length,d=Sn(t.call(p,p&&p.__data__,s,n)),h=d.length,x=f[s]=new Array(h),b=a[s]=new Array(h),g=l[s]=new Array(c);r(p,u,x,b,g,d,e);for(var y=0,v=0,w,C;y<h;++y)if(w=x[y]){for(y>=v&&(v=y+1);!(C=b[v])&&++v<h;);w._next=C||null}}return a=new $(a,n),a._enter=f,a._exit=l,a}function Sn(t){return typeof t=="object"&&"length"in t?t:Array.from(t)}function Qe(){return new $(this._exit||this._groups.map(Pt),this._parents)}function We(t,e,r){var n=this.enter(),o=this,i=this.exit();return typeof t=="function"?(n=t(n),n&&(n=n.selection())):n=n.append(t+""),e!=null&&(o=e(o),o&&(o=o.selection())),r==null?i.remove():r(i),n&&o?n.merge(o).order():o}function Ze(t){for(var e=t.selection?t.selection():t,r=this._groups,n=e._groups,o=r.length,i=n.length,a=Math.min(o,i),f=new Array(o),l=0;l<a;++l)for(var s=r[l],p=n[l],u=s.length,c=f[l]=new Array(u),d,h=0;h<u;++h)(d=s[h]||p[h])&&(c[h]=d);for(;l<o;++l)f[l]=r[l];return new $(f,this._parents)}function Ke(){for(var t=this._groups,e=-1,r=t.length;++e<r;)for(var n=t[e],o=n.length-1,i=n[o],a;--o>=0;)(a=n[o])&&(i&&a.compareDocumentPosition(i)^4&&i.parentNode.insertBefore(a,i),i=a);return this}function Je(t){t||(t=Nn);function e(u,c){return u&&c?t(u.__data__,c.__data__):!u-!c}for(var r=this._groups,n=r.length,o=new Array(n),i=0;i<n;++i){for(var a=r[i],f=a.length,l=o[i]=new Array(f),s,p=0;p<f;++p)(s=a[p])&&(l[p]=s);l.sort(e)}return new $(o,this._parents).order()}function Nn(t,e){return t<e?-1:t>e?1:t>=e?0:NaN}function je(){var t=arguments[0];return arguments[0]=this,t.apply(null,arguments),this}function tr(){return Array.from(this)}function er(){for(var t=this._groups,e=0,r=t.length;e<r;++e)for(var n=t[e],o=0,i=n.length;o<i;++o){var a=n[o];if(a)return a}return null}function rr(){let t=0;for(let e of this)++t;return t}function nr(){return!this.node()}function or(t){for(var e=this._groups,r=0,n=e.length;r<n;++r)for(var o=e[r],i=0,a=o.length,f;i<a;++i)(f=o[i])&&t.call(f,f.__data__,i,o);return this}function Rn(t){return function(){this.removeAttribute(t)}}function Cn(t){return function(){this.removeAttributeNS(t.space,t.local)}}function Pn(t,e){return function(){this.setAttribute(t,e)}}function Ln(t,e){return function(){this.setAttributeNS(t.space,t.local,e)}}function En(t,e){return function(){var r=e.apply(this,arguments);r==null?this.removeAttribute(t):this.setAttribute(t,r)}}function qn(t,e){return function(){var r=e.apply(this,arguments);r==null?this.removeAttributeNS(t.space,t.local):this.setAttributeNS(t.space,t.local,r)}}function ir(t,e){var r=St(t);if(arguments.length<2){var n=this.node();return r.local?n.getAttributeNS(r.space,r.local):n.getAttribute(r)}return this.each((e==null?r.local?Cn:Rn:typeof e=="function"?r.local?qn:En:r.local?Ln:Pn)(r,e))}function Lt(t){return t.ownerDocument&&t.ownerDocument.defaultView||t.document&&t||t.defaultView}function In(t){return function(){this.style.removeProperty(t)}}function On(t,e,r){return function(){this.style.setProperty(t,e,r)}}function Hn(t,e,r){return function(){var n=e.apply(this,arguments);n==null?this.style.removeProperty(t):this.style.setProperty(t,n,r)}}function ar(t,e,r){return arguments.length>1?this.each((e==null?In:typeof e=="function"?Hn:On)(t,e,r??"")):Bn(this.node(),t)}function Bn(t,e){return t.style.getPropertyValue(e)||Lt(t).getComputedStyle(t,null).getPropertyValue(e)}function Fn(t){return function(){delete this[t]}}function Dn(t,e){return function(){this[t]=e}}function Tn(t,e){return function(){var r=e.apply(this,arguments);r==null?delete this[t]:this[t]=r}}function fr(t,e){return arguments.length>1?this.each((e==null?Fn:typeof e=="function"?Tn:Dn)(t,e)):this.node()[t]}function sr(t){return t.trim().split(/^|\s+/)}function ne(t){return t.classList||new lr(t)}function lr(t){this._node=t,this._names=sr(t.getAttribute("class")||"")}lr.prototype={add:function(t){var e=this._names.indexOf(t);e<0&&(this._names.push(t),this._node.setAttribute("class",this._names.join(" ")))},remove:function(t){var e=this._names.indexOf(t);e>=0&&(this._names.splice(e,1),this._node.setAttribute("class",this._names.join(" ")))},contains:function(t){return this._names.indexOf(t)>=0}};function ur(t,e){for(var r=ne(t),n=-1,o=e.length;++n<o;)r.add(e[n])}function cr(t,e){for(var r=ne(t),n=-1,o=e.length;++n<o;)r.remove(e[n])}function zn(t){return function(){ur(this,t)}}function Vn(t){return function(){cr(this,t)}}function Xn(t,e){return function(){(e.apply(this,arguments)?ur:cr)(this,t)}}function pr(t,e){var r=sr(t+"");if(arguments.length<2){for(var n=ne(this.node()),o=-1,i=r.length;++o<i;)if(!n.contains(r[o]))return!1;return!0}return this.each((typeof e=="function"?Xn:e?zn:Vn)(r,e))}function Yn(){this.textContent=""}function Gn(t){return function(){this.textContent=t}}function Un(t){return function(){var e=t.apply(this,arguments);this.textContent=e??""}}function dr(t){return arguments.length?this.each(t==null?Yn:(typeof t=="function"?Un:Gn)(t)):this.node().textContent}function Qn(){this.innerHTML=""}function Wn(t){return function(){this.innerHTML=t}}function Zn(t){return function(){var e=t.apply(this,arguments);this.innerHTML=e??""}}function mr(t){return arguments.length?this.each(t==null?Qn:(typeof t=="function"?Zn:Wn)(t)):this.node().innerHTML}function Kn(){this.nextSibling&&this.parentNode.appendChild(this)}function hr(){return this.each(Kn)}function Jn(){this.previousSibling&&this.parentNode.insertBefore(this,this.parentNode.firstChild)}function xr(){return this.each(Jn)}function gr(t){var e=typeof t=="function"?t:Nt(t);return this.select(function(){return this.appendChild(e.apply(this,arguments))})}function jn(){return null}function yr(t,e){var r=typeof t=="function"?t:Nt(t),n=e==null?jn:typeof e=="function"?e:Rt(e);return this.select(function(){return this.insertBefore(r.apply(this,arguments),n.apply(this,arguments)||null)})}function to(){var t=this.parentNode;t&&t.removeChild(this)}function br(){return this.each(to)}function eo(){var t=this.cloneNode(!1),e=this.parentNode;return e?e.insertBefore(t,this.nextSibling):t}function ro(){var t=this.cloneNode(!0),e=this.parentNode;return e?e.insertBefore(t,this.nextSibling):t}function wr(t){return this.select(t?ro:eo)}function _r(t){return arguments.length?this.property("__data__",t):this.node().__data__}function no(t){return function(e){t.call(this,e,this.__data__)}}function oo(t){return t.trim().split(/^|\s+/).map(function(e){var r="",n=e.indexOf(".");return n>=0&&(r=e.slice(n+1),e=e.slice(0,n)),{type:e,name:r}})}function io(t){return function(){var e=this.__on;if(e){for(var r=0,n=-1,o=e.length,i;r<o;++r)i=e[r],(!t.type||i.type===t.type)&&i.name===t.name?this.removeEventListener(i.type,i.listener,i.options):e[++n]=i;++n?e.length=n:delete this.__on}}}function ao(t,e,r){return function(){var n=this.__on,o,i=no(e);if(n){for(var a=0,f=n.length;a<f;++a)if((o=n[a]).type===t.type&&o.name===t.name){this.removeEventListener(o.type,o.listener,o.options),this.addEventListener(o.type,o.listener=i,o.options=r),o.value=e;return}}this.addEventListener(t.type,i,r),o={type:t.type,name:t.name,value:e,listener:i,options:r},n?n.push(o):this.__on=[o]}}function vr(t,e,r){var n=oo(t+""),o,i=n.length,a;if(arguments.length<2){var f=this.node().__on;if(f){for(var l=0,s=f.length,p;l<s;++l)for(o=0,p=f[l];o<i;++o)if((a=n[o]).type===p.type&&a.name===p.name)return p.value}return}for(f=e?ao:io,o=0;o<i;++o)this.each(f(n[o],e,r));return this}function kr(t,e,r){var n=Lt(t),o=n.CustomEvent;typeof o=="function"?o=new o(e,r):(o=n.document.createEvent("Event"),r?(o.initEvent(e,r.bubbles,r.cancelable),o.detail=r.detail):o.initEvent(e,!1,!1)),t.dispatchEvent(o)}function fo(t,e){return function(){return kr(this,t,e)}}function so(t,e){return function(){return kr(this,t,e.apply(this,arguments))}}function Mr(t,e){return this.each((typeof e=="function"?so:fo)(t,e))}function*Ar(){for(var t=this._groups,e=0,r=t.length;e<r;++e)for(var n=t[e],o=0,i=n.length,a;o<i;++o)(a=n[o])&&(yield a)}var oe=[null];function $(t,e){this._groups=t,this._parents=e}function lo(){return new $([[document.documentElement]],oe)}function uo(){return this}$.prototype=lo.prototype={constructor:$,select:Be,selectAll:De,selectChild:ze,selectChildren:Ve,filter:Xe,data:Ue,enter:Ye,exit:Qe,join:We,merge:Ze,selection:uo,order:Ke,sort:Je,call:je,nodes:tr,node:er,size:rr,empty:nr,each:or,attr:ir,style:ar,property:fr,classed:pr,text:dr,html:mr,raise:hr,lower:xr,append:gr,insert:yr,remove:br,clone:wr,datum:_r,on:vr,dispatch:Mr,[Symbol.iterator]:Ar};function dt(t){return typeof t=="string"?new $([[document.querySelector(t)]],[document.documentElement]):new $([[t]],oe)}function W(t,e){return t==null||e==null?NaN:t<e?-1:t>e?1:t>=e?0:NaN}function ie(t,e){return t==null||e==null?NaN:e<t?-1:e>t?1:e>=t?0:NaN}function Et(t){let e,r,n;t.length!==2?(e=W,r=(f,l)=>W(t(f),l),n=(f,l)=>t(f)-l):(e=t===W||t===ie?t:co,r=t,n=t);function o(f,l,s=0,p=f.length){if(s<p){if(e(l,l)!==0)return p;do{let u=s+p>>>1;r(f[u],l)<0?s=u+1:p=u}while(s<p)}return s}function i(f,l,s=0,p=f.length){if(s<p){if(e(l,l)!==0)return p;do{let u=s+p>>>1;r(f[u],l)<=0?s=u+1:p=u}while(s<p)}return s}function a(f,l,s=0,p=f.length){let u=o(f,l,s,p-1);return u>s&&n(f[u-1],l)>-n(f[u],l)?u-1:u}return{left:o,center:a,right:i}}function co(){return 0}function ae(t){return t===null?NaN:+t}var $r=Et(W),Sr=$r.right,po=$r.left,mo=Et(ae).center,fe=Sr;var ot=class extends Map{constructor(e,r=go){if(super(),Object.defineProperties(this,{_intern:{value:new Map},_key:{value:r}}),e!=null)for(let[n,o]of e)this.set(n,o)}get(e){return super.get(Nr(this,e))}has(e){return super.has(Nr(this,e))}set(e,r){return super.set(ho(this,e),r)}delete(e){return super.delete(xo(this,e))}};function Nr({_intern:t,_key:e},r){let n=e(r);return t.has(n)?t.get(n):r}function ho({_intern:t,_key:e},r){let n=e(r);return t.has(n)?t.get(n):(t.set(n,r),r)}function xo({_intern:t,_key:e},r){let n=e(r);return t.has(n)&&(r=t.get(n),t.delete(n)),r}function go(t){return t!==null&&typeof t=="object"?t.valueOf():t}var yo=Math.sqrt(50),bo=Math.sqrt(10),wo=Math.sqrt(2);function qt(t,e,r){let n=(e-t)/Math.max(0,r),o=Math.floor(Math.log10(n)),i=n/Math.pow(10,o),a=i>=yo?10:i>=bo?5:i>=wo?2:1,f,l,s;return o<0?(s=Math.pow(10,-o)/a,f=Math.round(t*s),l=Math.round(e*s),f/s<t&&++f,l/s>e&&--l,s=-s):(s=Math.pow(10,o)*a,f=Math.round(t/s),l=Math.round(e/s),f*s<t&&++f,l*s>e&&--l),l<f&&.5<=r&&r<2?qt(t,e,r*2):[f,l,s]}function It(t,e,r){if(e=+e,t=+t,r=+r,!(r>0))return[];if(t===e)return[t];let n=e<t,[o,i,a]=n?qt(e,t,r):qt(t,e,r);if(!(i>=o))return[];let f=i-o+1,l=new Array(f);if(n)if(a<0)for(let s=0;s<f;++s)l[s]=(i-s)/-a;else for(let s=0;s<f;++s)l[s]=(i-s)*a;else if(a<0)for(let s=0;s<f;++s)l[s]=(o+s)/-a;else for(let s=0;s<f;++s)l[s]=(o+s)*a;return l}function mt(t,e,r){return e=+e,t=+t,r=+r,qt(t,e,r)[2]}function se(t,e,r){e=+e,t=+t,r=+r;let n=e<t,o=n?mt(e,t,r):mt(t,e,r);return(n?-1:1)*(o<0?1/-o:o)}function Ot(t,e){let r;if(e===void 0)for(let n of t)n!=null&&(r<n||r===void 0&&n>=n)&&(r=n);else{let n=-1;for(let o of t)(o=e(o,++n,t))!=null&&(r<o||r===void 0&&o>=o)&&(r=o)}return r}function Ht(t,e){let r;if(e===void 0)for(let n of t)n!=null&&(r>n||r===void 0&&n>=n)&&(r=n);else{let n=-1;for(let o of t)(o=e(o,++n,t))!=null&&(r>o||r===void 0&&o>=o)&&(r=o)}return r}function Bt(t,e,r){t=+t,e=+e,r=(o=arguments.length)<2?(e=t,t=0,1):o<3?1:+r;for(var n=-1,o=Math.max(0,Math.ceil((e-t)/r))|0,i=new Array(o);++n<o;)i[n]=t+n*r;return i}function it(t,e){switch(arguments.length){case 0:break;case 1:this.range(t);break;default:this.range(e).domain(t);break}return this}var Rr=Symbol("implicit");function Ft(){var t=new ot,e=[],r=[],n=Rr;function o(i){let a=t.get(i);if(a===void 0){if(n!==Rr)return n;t.set(i,a=e.push(i)-1)}return r[a%r.length]}return o.domain=function(i){if(!arguments.length)return e.slice();e=[],t=new ot;for(let a of i)t.has(a)||t.set(a,e.push(a)-1);return o},o.range=function(i){return arguments.length?(r=Array.from(i),o):r.slice()},o.unknown=function(i){return arguments.length?(n=i,o):n},o.copy=function(){return Ft(e,r).unknown(n)},it.apply(o,arguments),o}function ht(){var t=Ft().unknown(void 0),e=t.domain,r=t.range,n=0,o=1,i,a,f=!1,l=0,s=0,p=.5;delete t.unknown;function u(){var c=e().length,d=o<n,h=d?o:n,x=d?n:o;i=(x-h)/Math.max(1,c-l+s*2),f&&(i=Math.floor(i)),h+=(x-h-i*(c-l))*p,a=i*(1-l),f&&(h=Math.round(h),a=Math.round(a));var b=Bt(c).map(function(g){return h+i*g});return r(d?b.reverse():b)}return t.domain=function(c){return arguments.length?(e(c),u()):e()},t.range=function(c){return arguments.length?([n,o]=c,n=+n,o=+o,u()):[n,o]},t.rangeRound=function(c){return[n,o]=c,n=+n,o=+o,f=!0,u()},t.bandwidth=function(){return a},t.step=function(){return i},t.round=function(c){return arguments.length?(f=!!c,u()):f},t.padding=function(c){return arguments.length?(l=Math.min(1,s=+c),u()):l},t.paddingInner=function(c){return arguments.length?(l=Math.min(1,c),u()):l},t.paddingOuter=function(c){return arguments.length?(s=+c,u()):s},t.align=function(c){return arguments.length?(p=Math.max(0,Math.min(1,c)),u()):p},t.copy=function(){return ht(e(),[n,o]).round(f).paddingInner(l).paddingOuter(s).align(p)},it.apply(u(),arguments)}function Dt(t,e,r){t.prototype=e.prototype=r,r.constructor=t}function le(t,e){var r=Object.create(t.prototype);for(var n in e)r[n]=e[n];return r}function yt(){}var xt=.7,Vt=1/xt,at="\\s*([+-]?\\d+)\\s*",gt="\\s*([+-]?(?:\\d*\\.)?\\d+(?:[eE][+-]?\\d+)?)\\s*",D="\\s*([+-]?(?:\\d*\\.)?\\d+(?:[eE][+-]?\\d+)?)%\\s*",_o=/^#([0-9a-f]{3,8})$/,vo=new RegExp(`^rgb\\(${at},${at},${at}\\)$`),ko=new RegExp(`^rgb\\(${D},${D},${D}\\)$`),Mo=new RegExp(`^rgba\\(${at},${at},${at},${gt}\\)$`),Ao=new RegExp(`^rgba\\(${D},${D},${D},${gt}\\)$`),$o=new RegExp(`^hsl\\(${gt},${D},${D}\\)$`),So=new RegExp(`^hsla\\(${gt},${D},${D},${gt}\\)$`),Cr={aliceblue:15792383,antiquewhite:16444375,aqua:65535,aquamarine:8388564,azure:15794175,beige:16119260,bisque:16770244,black:0,blanchedalmond:16772045,blue:255,blueviolet:9055202,brown:10824234,burlywood:14596231,cadetblue:6266528,chartreuse:8388352,chocolate:13789470,coral:16744272,cornflowerblue:6591981,cornsilk:16775388,crimson:14423100,cyan:65535,darkblue:139,darkcyan:35723,darkgoldenrod:12092939,darkgray:11119017,darkgreen:25600,darkgrey:11119017,darkkhaki:12433259,darkmagenta:9109643,darkolivegreen:5597999,darkorange:16747520,darkorchid:10040012,darkred:9109504,darksalmon:15308410,darkseagreen:9419919,darkslateblue:4734347,darkslategray:3100495,darkslategrey:3100495,darkturquoise:52945,darkviolet:9699539,deeppink:16716947,deepskyblue:49151,dimgray:6908265,dimgrey:6908265,dodgerblue:2003199,firebrick:11674146,floralwhite:16775920,forestgreen:2263842,fuchsia:16711935,gainsboro:14474460,ghostwhite:16316671,gold:16766720,goldenrod:14329120,gray:8421504,green:32768,greenyellow:11403055,grey:8421504,honeydew:15794160,hotpink:16738740,indianred:13458524,indigo:4915330,ivory:16777200,khaki:15787660,lavender:15132410,lavenderblush:16773365,lawngreen:8190976,lemonchiffon:16775885,lightblue:11393254,lightcoral:15761536,lightcyan:14745599,lightgoldenrodyellow:16448210,lightgray:13882323,lightgreen:9498256,lightgrey:13882323,lightpink:16758465,lightsalmon:16752762,lightseagreen:2142890,lightskyblue:8900346,lightslategray:7833753,lightslategrey:7833753,lightsteelblue:11584734,lightyellow:16777184,lime:65280,limegreen:3329330,linen:16445670,magenta:16711935,maroon:8388608,mediumaquamarine:6737322,mediumblue:205,mediumorchid:12211667,mediumpurple:9662683,mediumseagreen:3978097,mediumslateblue:8087790,mediumspringgreen:64154,mediumturquoise:4772300,mediumvioletred:13047173,midnightblue:1644912,mintcream:16121850,mistyrose:16770273,moccasin:16770229,navajowhite:16768685,navy:128,oldlace:16643558,olive:8421376,olivedrab:7048739,orange:16753920,orangered:16729344,orchid:14315734,palegoldenrod:15657130,palegreen:10025880,paleturquoise:11529966,palevioletred:14381203,papayawhip:16773077,peachpuff:16767673,peru:13468991,pink:16761035,plum:14524637,powderblue:11591910,purple:8388736,rebeccapurple:6697881,red:16711680,rosybrown:12357519,royalblue:4286945,saddlebrown:9127187,salmon:16416882,sandybrown:16032864,seagreen:3050327,seashell:16774638,sienna:10506797,silver:12632256,skyblue:8900331,slateblue:6970061,slategray:7372944,slategrey:7372944,snow:16775930,springgreen:65407,steelblue:4620980,tan:13808780,teal:32896,thistle:14204888,tomato:16737095,turquoise:4251856,violet:15631086,wheat:16113331,white:16777215,whitesmoke:16119285,yellow:16776960,yellowgreen:10145074};Dt(yt,Y,{copy(t){return Object.assign(new this.constructor,this,t)},displayable(){return this.rgb().displayable()},hex:Pr,formatHex:Pr,formatHex8:No,formatHsl:Ro,formatRgb:Lr,toString:Lr});function Pr(){return this.rgb().formatHex()}function No(){return this.rgb().formatHex8()}function Ro(){return Br(this).formatHsl()}function Lr(){return this.rgb().formatRgb()}function Y(t){var e,r;return t=(t+"").trim().toLowerCase(),(e=_o.exec(t))?(r=e[1].length,e=parseInt(e[1],16),r===6?Er(e):r===3?new O(e>>8&15|e>>4&240,e>>4&15|e&240,(e&15)<<4|e&15,1):r===8?Tt(e>>24&255,e>>16&255,e>>8&255,(e&255)/255):r===4?Tt(e>>12&15|e>>8&240,e>>8&15|e>>4&240,e>>4&15|e&240,((e&15)<<4|e&15)/255):null):(e=vo.exec(t))?new O(e[1],e[2],e[3],1):(e=ko.exec(t))?new O(e[1]*255/100,e[2]*255/100,e[3]*255/100,1):(e=Mo.exec(t))?Tt(e[1],e[2],e[3],e[4]):(e=Ao.exec(t))?Tt(e[1]*255/100,e[2]*255/100,e[3]*255/100,e[4]):(e=$o.exec(t))?Or(e[1],e[2]/100,e[3]/100,1):(e=So.exec(t))?Or(e[1],e[2]/100,e[3]/100,e[4]):Cr.hasOwnProperty(t)?Er(Cr[t]):t==="transparent"?new O(NaN,NaN,NaN,0):null}function Er(t){return new O(t>>16&255,t>>8&255,t&255,1)}function Tt(t,e,r,n){return n<=0&&(t=e=r=NaN),new O(t,e,r,n)}function Co(t){return t instanceof yt||(t=Y(t)),t?(t=t.rgb(),new O(t.r,t.g,t.b,t.opacity)):new O}function ft(t,e,r,n){return arguments.length===1?Co(t):new O(t,e,r,n??1)}function O(t,e,r,n){this.r=+t,this.g=+e,this.b=+r,this.opacity=+n}Dt(O,ft,le(yt,{brighter(t){return t=t==null?Vt:Math.pow(Vt,t),new O(this.r*t,this.g*t,this.b*t,this.opacity)},darker(t){return t=t==null?xt:Math.pow(xt,t),new O(this.r*t,this.g*t,this.b*t,this.opacity)},rgb(){return this},clamp(){return new O(K(this.r),K(this.g),K(this.b),Xt(this.opacity))},displayable(){return-.5<=this.r&&this.r<255.5&&-.5<=this.g&&this.g<255.5&&-.5<=this.b&&this.b<255.5&&0<=this.opacity&&this.opacity<=1},hex:qr,formatHex:qr,formatHex8:Po,formatRgb:Ir,toString:Ir}));function qr(){return`#${Z(this.r)}${Z(this.g)}${Z(this.b)}`}function Po(){return`#${Z(this.r)}${Z(this.g)}${Z(this.b)}${Z((isNaN(this.opacity)?1:this.opacity)*255)}`}function Ir(){let t=Xt(this.opacity);return`${t===1?"rgb(":"rgba("}${K(this.r)}, ${K(this.g)}, ${K(this.b)}${t===1?")":`, ${t})`}`}function Xt(t){return isNaN(t)?1:Math.max(0,Math.min(1,t))}function K(t){return Math.max(0,Math.min(255,Math.round(t)||0))}function Z(t){return t=K(t),(t<16?"0":"")+t.toString(16)}function Or(t,e,r,n){return n<=0?t=e=r=NaN:r<=0||r>=1?t=e=NaN:e<=0&&(t=NaN),new B(t,e,r,n)}function Br(t){if(t instanceof B)return new B(t.h,t.s,t.l,t.opacity);if(t instanceof yt||(t=Y(t)),!t)return new B;if(t instanceof B)return t;t=t.rgb();var e=t.r/255,r=t.g/255,n=t.b/255,o=Math.min(e,r,n),i=Math.max(e,r,n),a=NaN,f=i-o,l=(i+o)/2;return f?(e===i?a=(r-n)/f+(r<n)*6:r===i?a=(n-e)/f+2:a=(e-r)/f+4,f/=l<.5?i+o:2-i-o,a*=60):f=l>0&&l<1?0:a,new B(a,f,l,t.opacity)}function Fr(t,e,r,n){return arguments.length===1?Br(t):new B(t,e,r,n??1)}function B(t,e,r,n){this.h=+t,this.s=+e,this.l=+r,this.opacity=+n}Dt(B,Fr,le(yt,{brighter(t){return t=t==null?Vt:Math.pow(Vt,t),new B(this.h,this.s,this.l*t,this.opacity)},darker(t){return t=t==null?xt:Math.pow(xt,t),new B(this.h,this.s,this.l*t,this.opacity)},rgb(){var t=this.h%360+(this.h<0)*360,e=isNaN(t)||isNaN(this.s)?0:this.s,r=this.l,n=r+(r<.5?r:1-r)*e,o=2*r-n;return new O(ue(t>=240?t-240:t+120,o,n),ue(t,o,n),ue(t<120?t+240:t-120,o,n),this.opacity)},clamp(){return new B(Hr(this.h),zt(this.s),zt(this.l),Xt(this.opacity))},displayable(){return(0<=this.s&&this.s<=1||isNaN(this.s))&&0<=this.l&&this.l<=1&&0<=this.opacity&&this.opacity<=1},formatHsl(){let t=Xt(this.opacity);return`${t===1?"hsl(":"hsla("}${Hr(this.h)}, ${zt(this.s)*100}%, ${zt(this.l)*100}%${t===1?")":`, ${t})`}`}}));function Hr(t){return t=(t||0)%360,t<0?t+360:t}function zt(t){return Math.max(0,Math.min(1,t||0))}function ue(t,e,r){return(t<60?e+(r-e)*t/60:t<180?r:t<240?e+(r-e)*(240-t)/60:e)*255}function ce(t,e,r,n,o){var i=t*t,a=i*t;return((1-3*t+3*i-a)*e+(4-6*i+3*a)*r+(1+3*t+3*i-3*a)*n+a*o)/6}function Dr(t){var e=t.length-1;return function(r){var n=r<=0?r=0:r>=1?(r=1,e-1):Math.floor(r*e),o=t[n],i=t[n+1],a=n>0?t[n-1]:2*o-i,f=n<e-1?t[n+2]:2*i-o;return ce((r-n/e)*e,a,o,i,f)}}function Tr(t){var e=t.length;return function(r){var n=Math.floor(((r%=1)<0?++r:r)*e),o=t[(n+e-1)%e],i=t[n%e],a=t[(n+1)%e],f=t[(n+2)%e];return ce((r-n/e)*e,o,i,a,f)}}var bt=t=>()=>t;function Lo(t,e){return function(r){return t+r*e}}function Eo(t,e,r){return t=Math.pow(t,r),e=Math.pow(e,r)-t,r=1/r,function(n){return Math.pow(t+n*e,r)}}function zr(t){return(t=+t)==1?Yt:function(e,r){return r-e?Eo(e,r,t):bt(isNaN(e)?r:e)}}function Yt(t,e){var r=e-t;return r?Lo(t,r):bt(isNaN(t)?e:t)}var pe=function t(e){var r=zr(e);function n(o,i){var a=r((o=ft(o)).r,(i=ft(i)).r),f=r(o.g,i.g),l=r(o.b,i.b),s=Yt(o.opacity,i.opacity);return function(p){return o.r=a(p),o.g=f(p),o.b=l(p),o.opacity=s(p),o+""}}return n.gamma=t,n}(1);function Vr(t){return function(e){var r=e.length,n=new Array(r),o=new Array(r),i=new Array(r),a,f;for(a=0;a<r;++a)f=ft(e[a]),n[a]=f.r||0,o[a]=f.g||0,i[a]=f.b||0;return n=t(n),o=t(o),i=t(i),f.opacity=1,function(l){return f.r=n(l),f.g=o(l),f.b=i(l),f+""}}}var Gf=Vr(Dr),Uf=Vr(Tr);function Xr(t,e){e||(e=[]);var r=t?Math.min(e.length,t.length):0,n=e.slice(),o;return function(i){for(o=0;o<r;++o)n[o]=t[o]*(1-i)+e[o]*i;return n}}function Yr(t){return ArrayBuffer.isView(t)&&!(t instanceof DataView)}function Gr(t,e){var r=e?e.length:0,n=t?Math.min(r,t.length):0,o=new Array(n),i=new Array(r),a;for(a=0;a<n;++a)o[a]=J(t[a],e[a]);for(;a<r;++a)i[a]=e[a];return function(f){for(a=0;a<n;++a)i[a]=o[a](f);return i}}function Ur(t,e){var r=new Date;return t=+t,e=+e,function(n){return r.setTime(t*(1-n)+e*n),r}}function G(t,e){return t=+t,e=+e,function(r){return t*(1-r)+e*r}}function Qr(t,e){var r={},n={},o;(t===null||typeof t!="object")&&(t={}),(e===null||typeof e!="object")&&(e={});for(o in e)o in t?r[o]=J(t[o],e[o]):n[o]=e[o];return function(i){for(o in r)n[o]=r[o](i);return n}}var me=/[-+]?(?:\d+\.?\d*|\.?\d+)(?:[eE][-+]?\d+)?/g,de=new RegExp(me.source,"g");function qo(t){return function(){return t}}function Io(t){return function(e){return t(e)+""}}function Wr(t,e){var r=me.lastIndex=de.lastIndex=0,n,o,i,a=-1,f=[],l=[];for(t=t+"",e=e+"";(n=me.exec(t))&&(o=de.exec(e));)(i=o.index)>r&&(i=e.slice(r,i),f[a]?f[a]+=i:f[++a]=i),(n=n[0])===(o=o[0])?f[a]?f[a]+=o:f[++a]=o:(f[++a]=null,l.push({i:a,x:G(n,o)})),r=de.lastIndex;return r<e.length&&(i=e.slice(r),f[a]?f[a]+=i:f[++a]=i),f.length<2?l[0]?Io(l[0].x):qo(e):(e=l.length,function(s){for(var p=0,u;p<e;++p)f[(u=l[p]).i]=u.x(s);return f.join("")})}function J(t,e){var r=typeof e,n;return e==null||r==="boolean"?bt(e):(r==="number"?G:r==="string"?(n=Y(e))?(e=n,pe):Wr:e instanceof Y?pe:e instanceof Date?Ur:Yr(e)?Xr:Array.isArray(e)?Gr:typeof e.valueOf!="function"&&typeof e.toString!="function"||isNaN(e)?Qr:G)(t,e)}function he(t,e){return t=+t,e=+e,function(r){return Math.round(t*(1-r)+e*r)}}function xe(t){return function(){return t}}function ge(t){return+t}var Zr=[0,1];function st(t){return t}function ye(t,e){return(e-=t=+t)?function(r){return(r-t)/e}:xe(isNaN(e)?NaN:.5)}function Oo(t,e){var r;return t>e&&(r=t,t=e,e=r),function(n){return Math.max(t,Math.min(e,n))}}function Ho(t,e,r){var n=t[0],o=t[1],i=e[0],a=e[1];return o<n?(n=ye(o,n),i=r(a,i)):(n=ye(n,o),i=r(i,a)),function(f){return i(n(f))}}function Bo(t,e,r){var n=Math.min(t.length,e.length)-1,o=new Array(n),i=new Array(n),a=-1;for(t[n]<t[0]&&(t=t.slice().reverse(),e=e.slice().reverse());++a<n;)o[a]=ye(t[a],t[a+1]),i[a]=r(e[a],e[a+1]);return function(f){var l=fe(t,f,1,n)-1;return i[l](o[l](f))}}function Kr(t,e){return e.domain(t.domain()).range(t.range()).interpolate(t.interpolate()).clamp(t.clamp()).unknown(t.unknown())}function Fo(){var t=Zr,e=Zr,r=J,n,o,i,a=st,f,l,s;function p(){var c=Math.min(t.length,e.length);return a!==st&&(a=Oo(t[0],t[c-1])),f=c>2?Bo:Ho,l=s=null,u}function u(c){return c==null||isNaN(c=+c)?i:(l||(l=f(t.map(n),e,r)))(n(a(c)))}return u.invert=function(c){return a(o((s||(s=f(e,t.map(n),G)))(c)))},u.domain=function(c){return arguments.length?(t=Array.from(c,ge),p()):t.slice()},u.range=function(c){return arguments.length?(e=Array.from(c),p()):e.slice()},u.rangeRound=function(c){return e=Array.from(c),r=he,p()},u.clamp=function(c){return arguments.length?(a=c?!0:st,p()):a!==st},u.interpolate=function(c){return arguments.length?(r=c,p()):r},u.unknown=function(c){return arguments.length?(i=c,u):i},function(c,d){return n=c,o=d,p()}}function be(){return Fo()(st,st)}function Jr(t){return Math.abs(t=Math.round(t))>=1e21?t.toLocaleString("en").replace(/,/g,""):t.toString(10)}function j(t,e){if(!isFinite(t)||t===0)return null;var r=(t=e?t.toExponential(e-1):t.toExponential()).indexOf("e"),n=t.slice(0,r);return[n.length>1?n[0]+n.slice(2):n,+t.slice(r+1)]}function T(t){return t=j(Math.abs(t)),t?t[1]:NaN}function jr(t,e){return function(r,n){for(var o=r.length,i=[],a=0,f=t[0],l=0;o>0&&f>0&&(l+f+1>n&&(f=Math.max(1,n-l)),i.push(r.substring(o-=f,o+f)),!((l+=f+1)>n));)f=t[a=(a+1)%t.length];return i.reverse().join(e)}}function tn(t){return function(e){return e.replace(/[0-9]/g,function(r){return t[+r]})}}var Do=/^(?:(.)?([<>=^]))?([+\-( ])?([$#])?(0)?(\d+)?(,)?(\.\d+)?(~)?([a-z%])?$/i;function U(t){if(!(e=Do.exec(t)))throw new Error("invalid format: "+t);var e;return new Gt({fill:e[1],align:e[2],sign:e[3],symbol:e[4],zero:e[5],width:e[6],comma:e[7],precision:e[8]&&e[8].slice(1),trim:e[9],type:e[10]})}U.prototype=Gt.prototype;function Gt(t){this.fill=t.fill===void 0?" ":t.fill+"",this.align=t.align===void 0?">":t.align+"",this.sign=t.sign===void 0?"-":t.sign+"",this.symbol=t.symbol===void 0?"":t.symbol+"",this.zero=!!t.zero,this.width=t.width===void 0?void 0:+t.width,this.comma=!!t.comma,this.precision=t.precision===void 0?void 0:+t.precision,this.trim=!!t.trim,this.type=t.type===void 0?"":t.type+""}Gt.prototype.toString=function(){return this.fill+this.align+this.sign+this.symbol+(this.zero?"0":"")+(this.width===void 0?"":Math.max(1,this.width|0))+(this.comma?",":"")+(this.precision===void 0?"":"."+Math.max(0,this.precision|0))+(this.trim?"~":"")+this.type};function en(t){t:for(var e=t.length,r=1,n=-1,o;r<e;++r)switch(t[r]){case".":n=o=r;break;case"0":n===0&&(n=r),o=r;break;default:if(!+t[r])break t;n>0&&(n=0);break}return n>0?t.slice(0,n)+t.slice(o+1):t}var wt;function rn(t,e){var r=j(t,e);if(!r)return wt=void 0,t.toPrecision(e);var n=r[0],o=r[1],i=o-(wt=Math.max(-8,Math.min(8,Math.floor(o/3)))*3)+1,a=n.length;return i===a?n:i>a?n+new Array(i-a+1).join("0"):i>0?n.slice(0,i)+"."+n.slice(i):"0."+new Array(1-i).join("0")+j(t,Math.max(0,e+i-1))[0]}function we(t,e){var r=j(t,e);if(!r)return t+"";var n=r[0],o=r[1];return o<0?"0."+new Array(-o).join("0")+n:n.length>o+1?n.slice(0,o+1)+"."+n.slice(o+1):n+new Array(o-n.length+2).join("0")}var _e={"%":(t,e)=>(t*100).toFixed(e),b:t=>Math.round(t).toString(2),c:t=>t+"",d:Jr,e:(t,e)=>t.toExponential(e),f:(t,e)=>t.toFixed(e),g:(t,e)=>t.toPrecision(e),o:t=>Math.round(t).toString(8),p:(t,e)=>we(t*100,e),r:we,s:rn,X:t=>Math.round(t).toString(16).toUpperCase(),x:t=>Math.round(t).toString(16)};function ve(t){return t}var nn=Array.prototype.map,on=["y","z","a","f","p","n","\xB5","m","","k","M","G","T","P","E","Z","Y"];function an(t){var e=t.grouping===void 0||t.thousands===void 0?ve:jr(nn.call(t.grouping,Number),t.thousands+""),r=t.currency===void 0?"":t.currency[0]+"",n=t.currency===void 0?"":t.currency[1]+"",o=t.decimal===void 0?".":t.decimal+"",i=t.numerals===void 0?ve:tn(nn.call(t.numerals,String)),a=t.percent===void 0?"%":t.percent+"",f=t.minus===void 0?"\u2212":t.minus+"",l=t.nan===void 0?"NaN":t.nan+"";function s(u,c){u=U(u);var d=u.fill,h=u.align,x=u.sign,b=u.symbol,g=u.zero,y=u.width,v=u.comma,w=u.precision,C=u.trim,M=u.type;M==="n"?(v=!0,M="g"):_e[M]||(w===void 0&&(w=12),C=!0,M="g"),(g||d==="0"&&h==="=")&&(g=!0,d="0",h="=");var R=(c&&c.prefix!==void 0?c.prefix:"")+(b==="$"?r:b==="#"&&/[boxX]/.test(M)?"0"+M.toLowerCase():""),H=(b==="$"?n:/[%p]/.test(M)?a:"")+(c&&c.suffix!==void 0?c.suffix:""),z=_e[M],V=/[defgprs%]/.test(M);w=w===void 0?6:/[gprs]/.test(M)?Math.max(1,Math.min(21,w)):Math.max(0,Math.min(20,w));function F(m){var E=R,q=H,X,rt,A;if(M==="c")q=z(m)+q,m="";else{m=+m;var S=m<0||1/m<0;if(m=isNaN(m)?l:z(Math.abs(m),w),C&&(m=en(m)),S&&+m==0&&x!=="+"&&(S=!1),E=(S?x==="("?x:f:x==="-"||x==="("?"":x)+E,q=(M==="s"&&!isNaN(m)&&wt!==void 0?on[8+wt/3]:"")+q+(S&&x==="("?")":""),V){for(X=-1,rt=m.length;++X<rt;)if(A=m.charCodeAt(X),48>A||A>57){q=(A===46?o+m.slice(X+1):m.slice(X))+q,m=m.slice(0,X);break}}}v&&!g&&(m=e(m,1/0));var k=E.length+m.length+q.length,P=k<y?new Array(y-k+1).join(d):"";switch(v&&g&&(m=e(P+m,P.length?y-q.length:1/0),P=""),h){case"<":m=E+m+q+P;break;case"=":m=E+P+m+q;break;case"^":m=P.slice(0,k=P.length>>1)+E+m+q+P.slice(k);break;default:m=P+E+m+q;break}return i(m)}return F.toString=function(){return u+""},F}function p(u,c){var d=Math.max(-8,Math.min(8,Math.floor(T(c)/3)))*3,h=Math.pow(10,-d),x=s((u=U(u),u.type="f",u),{suffix:on[8+d/3]});return function(b){return x(h*b)}}return{format:s,formatPrefix:p}}var Ut,Qt,Wt;ke({thousands:",",grouping:[3],currency:["$",""]});function ke(t){return Ut=an(t),Qt=Ut.format,Wt=Ut.formatPrefix,Ut}function Me(t){return Math.max(0,-T(Math.abs(t)))}function Ae(t,e){return Math.max(0,Math.max(-8,Math.min(8,Math.floor(T(e)/3)))*3-T(Math.abs(t)))}function $e(t,e){return t=Math.abs(t),e=Math.abs(e)-t,Math.max(0,T(e)-T(t))+1}function Se(t,e,r,n){var o=se(t,e,r),i;switch(n=U(n??",f"),n.type){case"s":{var a=Math.max(Math.abs(t),Math.abs(e));return n.precision==null&&!isNaN(i=Ae(o,a))&&(n.precision=i),Wt(n,a)}case"":case"e":case"g":case"p":case"r":{n.precision==null&&!isNaN(i=$e(o,Math.max(Math.abs(t),Math.abs(e))))&&(n.precision=i-(n.type==="e"));break}case"f":case"%":{n.precision==null&&!isNaN(i=Me(o))&&(n.precision=i-(n.type==="%")*2);break}}return Qt(n)}function To(t){var e=t.domain;return t.ticks=function(r){var n=e();return It(n[0],n[n.length-1],r??10)},t.tickFormat=function(r,n){var o=e();return Se(o[0],o[o.length-1],r??10,n)},t.nice=function(r){r==null&&(r=10);var n=e(),o=0,i=n.length-1,a=n[o],f=n[i],l,s,p=10;for(f<a&&(s=a,a=f,f=s,s=o,o=i,i=s);p-- >0;){if(s=mt(a,f,r),s===l)return n[o]=a,n[i]=f,e(n);if(s>0)a=Math.floor(a/s)*s,f=Math.ceil(f/s)*s;else if(s<0)a=Math.ceil(a*s)/s,f=Math.floor(f*s)/s;else break;l=s}return t},t}function lt(){var t=be();return t.copy=function(){return Kr(t,lt())},it.apply(t,arguments),To(t)}function fn(t){return t}var Ne=1,Re=2,Ce=3,_t=4,sn=1e-6;function zo(t){return"translate("+t+",0)"}function Vo(t){return"translate(0,"+t+")"}function Xo(t){return e=>+t(e)}function Yo(t,e){return e=Math.max(0,t.bandwidth()-e*2)/2,t.round()&&(e=Math.round(e)),r=>+t(r)+e}function Go(){return!this.__axis}function ln(t,e){var r=[],n=null,o=null,i=6,a=6,f=3,l=typeof window<"u"&&window.devicePixelRatio>1?0:.5,s=t===Ne||t===_t?-1:1,p=t===_t||t===Re?"x":"y",u=t===Ne||t===Ce?zo:Vo;function c(d){var h=n??(e.ticks?e.ticks.apply(e,r):e.domain()),x=o??(e.tickFormat?e.tickFormat.apply(e,r):fn),b=Math.max(i,0)+f,g=e.range(),y=+g[0]+l,v=+g[g.length-1]+l,w=(e.bandwidth?Yo:Xo)(e.copy(),l),C=d.selection?d.selection():d,M=C.selectAll(".domain").data([null]),R=C.selectAll(".tick").data(h,e).order(),H=R.exit(),z=R.enter().append("g").attr("class","tick"),V=R.select("line"),F=R.select("text");M=M.merge(M.enter().insert("path",".tick").attr("class","domain").attr("stroke","currentColor")),R=R.merge(z),V=V.merge(z.append("line").attr("stroke","currentColor").attr(p+"2",s*i)),F=F.merge(z.append("text").attr("fill","currentColor").attr(p,s*b).attr("dy",t===Ne?"0em":t===Ce?"0.71em":"0.32em")),d!==C&&(M=M.transition(d),R=R.transition(d),V=V.transition(d),F=F.transition(d),H=H.transition(d).attr("opacity",sn).attr("transform",function(m){return isFinite(m=w(m))?u(m+l):this.getAttribute("transform")}),z.attr("opacity",sn).attr("transform",function(m){var E=this.parentNode.__axis;return u((E&&isFinite(E=E(m))?E:w(m))+l)})),H.remove(),M.attr("d",t===_t||t===Re?a?"M"+s*a+","+y+"H"+l+"V"+v+"H"+s*a:"M"+l+","+y+"V"+v:a?"M"+y+","+s*a+"V"+l+"H"+v+"V"+s*a:"M"+y+","+l+"H"+v),R.attr("opacity",1).attr("transform",function(m){return u(w(m)+l)}),V.attr(p+"2",s*i),F.attr(p,s*b).text(x),C.filter(Go).attr("fill","none").attr("font-size",10).attr("font-family","sans-serif").attr("text-anchor",t===Re?"start":t===_t?"end":"middle"),C.each(function(){this.__axis=w})}return c.scale=function(d){return arguments.length?(e=d,c):e},c.ticks=function(){return r=Array.from(arguments),c},c.tickArguments=function(d){return arguments.length?(r=d==null?[]:Array.from(d),c):r.slice()},c.tickValues=function(d){return arguments.length?(n=d==null?null:Array.from(d),c):n&&n.slice()},c.tickFormat=function(d){return arguments.length?(o=d,c):o},c.tickSize=function(d){return arguments.length?(i=a=+d,c):i},c.tickSizeInner=function(d){return arguments.length?(i=+d,c):i},c.tickSizeOuter=function(d){return arguments.length?(a=+d,c):a},c.tickPadding=function(d){return arguments.length?(f=+d,c):f},c.offset=function(d){return arguments.length?(l=+d,c):l},c}function Pe(t){return ln(Ce,t)}function Le(t){return ln(_t,t)}function N(t){return function(){return t}}var Ee=Math.PI,qe=2*Ee,tt=1e-6,Uo=qe-tt;function un(t){this._+=t[0];for(let e=1,r=t.length;e<r;++e)this._+=arguments[e]+t[e]}function Qo(t){let e=Math.floor(t);if(!(e>=0))throw new Error(`invalid digits: ${t}`);if(e>15)return un;let r=10**e;return function(n){this._+=n[0];for(let o=1,i=n.length;o<i;++o)this._+=Math.round(arguments[o]*r)/r+n[o]}}var et=class{constructor(e){this._x0=this._y0=this._x1=this._y1=null,this._="",this._append=e==null?un:Qo(e)}moveTo(e,r){this._append`M${this._x0=this._x1=+e},${this._y0=this._y1=+r}`}closePath(){this._x1!==null&&(this._x1=this._x0,this._y1=this._y0,this._append`Z`)}lineTo(e,r){this._append`L${this._x1=+e},${this._y1=+r}`}quadraticCurveTo(e,r,n,o){this._append`Q${+e},${+r},${this._x1=+n},${this._y1=+o}`}bezierCurveTo(e,r,n,o,i,a){this._append`C${+e},${+r},${+n},${+o},${this._x1=+i},${this._y1=+a}`}arcTo(e,r,n,o,i){if(e=+e,r=+r,n=+n,o=+o,i=+i,i<0)throw new Error(`negative radius: ${i}`);let a=this._x1,f=this._y1,l=n-e,s=o-r,p=a-e,u=f-r,c=p*p+u*u;if(this._x1===null)this._append`M${this._x1=e},${this._y1=r}`;else if(c>tt)if(!(Math.abs(u*l-s*p)>tt)||!i)this._append`L${this._x1=e},${this._y1=r}`;else{let d=n-a,h=o-f,x=l*l+s*s,b=d*d+h*h,g=Math.sqrt(x),y=Math.sqrt(c),v=i*Math.tan((Ee-Math.acos((x+c-b)/(2*g*y)))/2),w=v/y,C=v/g;Math.abs(w-1)>tt&&this._append`L${e+w*p},${r+w*u}`,this._append`A${i},${i},0,0,${+(u*d>p*h)},${this._x1=e+C*l},${this._y1=r+C*s}`}}arc(e,r,n,o,i,a){if(e=+e,r=+r,n=+n,a=!!a,n<0)throw new Error(`negative radius: ${n}`);let f=n*Math.cos(o),l=n*Math.sin(o),s=e+f,p=r+l,u=1^a,c=a?o-i:i-o;this._x1===null?this._append`M${s},${p}`:(Math.abs(this._x1-s)>tt||Math.abs(this._y1-p)>tt)&&this._append`L${s},${p}`,n&&(c<0&&(c=c%qe+qe),c>Uo?this._append`A${n},${n},0,1,${u},${e-f},${r-l}A${n},${n},0,1,${u},${this._x1=s},${this._y1=p}`:c>tt&&this._append`A${n},${n},0,${+(c>=Ee)},${u},${this._x1=e+n*Math.cos(i)},${this._y1=r+n*Math.sin(i)}`)}rect(e,r,n,o){this._append`M${this._x0=this._x1=+e},${this._y0=this._y1=+r}h${n=+n}v${+o}h${-n}Z`}toString(){return this._}};function cn(){return new et}cn.prototype=et.prototype;function Zt(t){let e=3;return t.digits=function(r){if(!arguments.length)return e;if(r==null)e=null;else{let n=Math.floor(r);if(!(n>=0))throw new RangeError(`invalid digits: ${r}`);e=n}return t},()=>new et(e)}var Hl=Array.prototype.slice;function Kt(t){return typeof t=="object"&&"length"in t?t:Array.from(t)}function pn(t){this._context=t}pn.prototype={areaStart:function(){this._line=0},areaEnd:function(){this._line=NaN},lineStart:function(){this._point=0},lineEnd:function(){(this._line||this._line!==0&&this._point===1)&&this._context.closePath(),this._line=1-this._line},point:function(t,e){switch(t=+t,e=+e,this._point){case 0:this._point=1,this._line?this._context.lineTo(t,e):this._context.moveTo(t,e);break;case 1:this._point=2;default:this._context.lineTo(t,e);break}}};function Jt(t){return new pn(t)}function jt(t){return t[0]}function te(t){return t[1]}function vt(t,e){var r=N(!0),n=null,o=Jt,i=null,a=Zt(f);t=typeof t=="function"?t:t===void 0?jt:N(t),e=typeof e=="function"?e:e===void 0?te:N(e);function f(l){var s,p=(l=Kt(l)).length,u,c=!1,d;for(n==null&&(i=o(d=a())),s=0;s<=p;++s)!(s<p&&r(u=l[s],s,l))===c&&((c=!c)?i.lineStart():i.lineEnd()),c&&i.point(+t(u,s,l),+e(u,s,l));if(d)return i=null,d+""||null}return f.x=function(l){return arguments.length?(t=typeof l=="function"?l:N(+l),f):t},f.y=function(l){return arguments.length?(e=typeof l=="function"?l:N(+l),f):e},f.defined=function(l){return arguments.length?(r=typeof l=="function"?l:N(!!l),f):r},f.curve=function(l){return arguments.length?(o=l,n!=null&&(i=o(n)),f):o},f.context=function(l){return arguments.length?(l==null?n=i=null:i=o(n=l),f):n},f}function Ie(t,e,r){var n=null,o=N(!0),i=null,a=Jt,f=null,l=Zt(s);t=typeof t=="function"?t:t===void 0?jt:N(+t),e=typeof e=="function"?e:e===void 0?N(0):N(+e),r=typeof r=="function"?r:r===void 0?te:N(+r);function s(u){var c,d,h,x=(u=Kt(u)).length,b,g=!1,y,v=new Array(x),w=new Array(x);for(i==null&&(f=a(y=l())),c=0;c<=x;++c){if(!(c<x&&o(b=u[c],c,u))===g)if(g=!g)d=c,f.areaStart(),f.lineStart();else{for(f.lineEnd(),f.lineStart(),h=c-1;h>=d;--h)f.point(v[h],w[h]);f.lineEnd(),f.areaEnd()}g&&(v[c]=+t(b,c,u),w[c]=+e(b,c,u),f.point(n?+n(b,c,u):v[c],r?+r(b,c,u):w[c]))}if(y)return f=null,y+""||null}function p(){return vt().defined(o).curve(a).context(i)}return s.x=function(u){return arguments.length?(t=typeof u=="function"?u:N(+u),n=null,s):t},s.x0=function(u){return arguments.length?(t=typeof u=="function"?u:N(+u),s):t},s.x1=function(u){return arguments.length?(n=u==null?null:typeof u=="function"?u:N(+u),s):n},s.y=function(u){return arguments.length?(e=typeof u=="function"?u:N(+u),r=null,s):e},s.y0=function(u){return arguments.length?(e=typeof u=="function"?u:N(+u),s):e},s.y1=function(u){return arguments.length?(r=u==null?null:typeof u=="function"?u:N(+u),s):r},s.lineX0=s.lineY0=function(){return p().x(t).y(e)},s.lineY1=function(){return p().x(t).y(r)},s.lineX1=function(){return p().x(n).y(e)},s.defined=function(u){return arguments.length?(o=typeof u=="function"?u:N(!!u),s):o},s.curve=function(u){return arguments.length?(a=u,i!=null&&(f=a(i)),s):a},s.context=function(u){return arguments.length?(u==null?i=f=null:f=a(i=u),s):i},s}function Wo({model:t,el:e}){let r=document.createElement("div");r.classList.add("pulsar-chart-container"),e.appendChild(r);let n=t.get("data")||[],o=t.get("x_values")||[],i=t.get("width"),a=t.get("height"),f=t.get("overlap"),l=t.get("stroke_width"),s=t.get("fill_opacity"),p=t.get("peak_scale")||1,u=t.get("selected_index"),c=t.get("x_label")||"",d=t.get("y_label")||"";function h(A){return A.map((S,k)=>({x:o[k]!==void 0?o[k]:k,y:S}))}let x={top:20,right:20,bottom:40,left:60},b=i-x.left-x.right,g=a-x.top-x.bottom,y=dt(r).append("svg").attr("width",i).attr("height",a).attr("class","pulsar-chart-svg"),v=y.append("g").attr("transform",`translate(${x.left}, ${x.top})`),w=`pulsar-clip-${Math.random().toString(36).substr(2,9)}`,C=y.append("defs").append("clipPath").attr("id",w).append("rect").attr("x",0).attr("y",0).attr("width",b).attr("height",g),M=v.append("g").attr("class","pulsar-rows").attr("clip-path",`url(#${w})`),R=v.append("g").attr("class","pulsar-y-axis"),H=v.append("g").attr("class","pulsar-x-axis").attr("transform",`translate(0, ${g})`),z=y.append("text").attr("class","pulsar-axis-label pulsar-x-label").attr("x",x.left+b/2).attr("y",a-5).attr("text-anchor","middle").text(c),V=y.append("text").attr("class","pulsar-axis-label pulsar-y-label").attr("transform","rotate(-90)").attr("x",-(x.top+g/2)).attr("y",15).attr("text-anchor","middle").text(d);function F(){let A=n.length;if(A===0)return{x:null,y:null,yBand:null};let S=Math.max(...n.map(_=>_.values.length)),k=lt().domain([0,S-1]).range([0,b]),P=n.map(_=>_.index),ut=ht().domain(P).range([g,0]).padding(0),kt=g/A*(1+f),nt=n.flatMap(_=>_.values),Mt=Ht(nt)||0,Q=Ot(nt)||1,ct=Q-Mt||1,At=kt*.4*p,He=lt().domain([Mt-ct*.1,Q+ct*.1]).range([At,-At]);return{x:k,yBand:ut,yAmplitude:He,rowHeight:kt,numRows:A}}function m(){let A=F();if(!A.x){M.selectAll("*").remove(),R.selectAll("*").remove(),H.selectAll("*").remove();return}let{x:S,yBand:k,yAmplitude:P,rowHeight:ut}=A;C.attr("height",g+ut);let Oe=P(0),kt=Le(k).tickSize(0).tickPadding(10);R.call(kt),R.select(".domain").remove(),R.selectAll(".tick").attr("transform",_=>`translate(0, ${k(_)+k.bandwidth()/2+Oe})`);let nt=ut*.5;H.attr("transform",`translate(0, ${g+nt})`),H.selectAll("*").remove();let Mt=Pe(S).ticks(10);H.call(Mt),H.select(".domain").remove(),z.attr("y",x.top+g+nt+35),y.attr("height",a+nt);let Q=vt().defined(_=>!isNaN(_)).x((_,L)=>S(L)).y(_=>P(_)),ct=Ie().defined(_=>!isNaN(_)).x((_,L)=>S(L)).y0(ut*.5).y1(_=>P(_)),At=[...n].reverse();M.selectAll(".pulsar-row").data(At,_=>_.index).join(_=>{let L=_.append("g").attr("class","pulsar-row").attr("transform",I=>`translate(0, ${k(I.index)+k.bandwidth()/2})`);return L.append("path").attr("class","pulsar-area").attr("d",I=>ct(I.values)).style("opacity",s),L.append("path").attr("class","pulsar-line-hitbox").attr("d",I=>Q(I.values)),L.append("path").attr("class","pulsar-line").attr("d",I=>Q(I.values)).style("stroke-width",l),L},_=>_.attr("transform",L=>`translate(0, ${k(L.index)+k.bandwidth()/2})`).call(L=>{L.select(".pulsar-area").attr("d",I=>ct(I.values)).style("opacity",s),L.select(".pulsar-line-hitbox").attr("d",I=>Q(I.values)),L.select(".pulsar-line").attr("d",I=>Q(I.values)).style("stroke-width",l)}),_=>_.remove()).selectAll(".pulsar-line-hitbox").on("mouseenter",E).on("mouseleave",q).on("click",X),rt()}function E(A,S){dt(this.parentNode).classed("pulsar-row-hover",!0),R.selectAll(".tick text").classed("pulsar-y-label-hover",k=>k===S.index)}function q(A,S){dt(this.parentNode).classed("pulsar-row-hover",!1),R.selectAll(".tick text").classed("pulsar-y-label-hover",!1)}function X(A,S){let k=S.index;u===k?(u=null,t.set("selected_index",null),t.set("selected_row",[])):(u=k,t.set("selected_index",k),t.set("selected_row",h(S.values))),t.save_changes(),rt()}function rt(){M.selectAll(".pulsar-row").classed("pulsar-row-selected",A=>A.index===u),R.selectAll(".tick text").classed("pulsar-y-label-selected",A=>A===u)}m(),t.on("change:data",()=>{n=t.get("data")||[],m()}),t.on("change:overlap",()=>{f=t.get("overlap"),m()}),t.on("change:stroke_width",()=>{l=t.get("stroke_width"),m()}),t.on("change:fill_opacity",()=>{s=t.get("fill_opacity"),m()}),t.on("change:peak_scale",()=>{p=t.get("peak_scale")||1,m()}),t.on("change:selected_index",()=>{if(u=t.get("selected_index"),u!==null){let A=n.find(S=>S.index===u);A&&(t.set("selected_row",h(A.values)),t.save_changes())}rt()}),t.on("change:x_values",()=>{o=t.get("x_values")||[]})}var su={render:Wo};export{su as default};
