+++ wigglystuff/static/ridgeline-chart.css
@@ -1,0 +1,131 @@
+/* RidgelineChart Widget Styles */
+.ridgeline-chart-container {
+    --ridgeline-bg: #000000;
+    --ridgeline-stroke: #ffffff;
+    --ridgeline-fill: #000000;
+    --ridgeline-hover-stroke: #ffcc00;
+    --ridgeline-selected-stroke: #ff6b6b;
+    --ridgeline-axis-color: #666666;
+    --ridgeline-label-color: #999999;
+    color-scheme: light dark;
+}
+
+/* Light mode overrides */
+@media (prefers-color-scheme: light) {
+    .ridgeline-chart-container {
+        --ridgeline-bg: #ffffff;
+        --ridgeline-stroke: #1a1a1a;
+        --ridgeline-fill: #ffffff;
+        --ridgeline-hover-stroke: #0066cc;
+        --ridgeline-selected-stroke: #ff6b6b;
+        --ridgeline-axis-color: #999999;
+        --ridgeline-label-color: #666666;
+    }
+}
+
+/* Explicit light theme selectors */
+:where(:not(.dark):not(.dark-theme):not([data-theme="dark"])) .ridgeline-chart-container {
+    --ridgeline-bg: #ffffff;
+    --ridgeline-stroke: #1a1a1a;
+    --ridgeline-fill: #ffffff;
+    --ridgeline-hover-stroke: #0066cc;
+    --ridgeline-selected-stroke: #ff6b6b;
+    --ridgeline-axis-color: #999999;
+    --ridgeline-label-color: #666666;
+}
+
+/* Dark mode - explicit selectors for marimo and jupyter */
+:where(.dark, .dark-theme, [data-theme="dark"], body.dark, body[data-theme="dark"]) .ridgeline-chart-container {
+    --ridgeline-bg: #000000;
+    --ridgeline-stroke: #ffffff;
+    --ridgeline-fill: #000000;
+    --ridgeline-hover-stroke: #ffcc00;
+    --ridgeline-selected-stroke: #ff6b6b;
+    --ridgeline-axis-color: #666666;
+    --ridgeline-label-color: #999999;
+}
+
+.ridgeline-chart-svg {
+    background-color: var(--ridgeline-bg);
+    display: block;
+}
+
+.ridgeline-row {
+    cursor: pointer;
+}
+
+.ridgeline-area {
+    fill: var(--ridgeline-fill);
+    stroke: none;
+    pointer-events: none;
+}
+
+/* Invisible thick stroke for easier click targeting */
+.ridgeline-line-hitbox {
+    fill: none;
+    stroke: transparent;
+    stroke-width: 12px;
+    stroke-linejoin: round;
+    stroke-linecap: round;
+    pointer-events: stroke;
+    cursor: pointer;
+}
+
+.ridgeline-line {
+    fill: none;
+    stroke: var(--ridgeline-stroke);
+    stroke-linejoin: round;
+    stroke-linecap: round;
+    pointer-events: none;
+}
+
+/* Hover state - thicker line, no color change */
+.ridgeline-row-hover .ridgeline-line {
+    stroke-width: 2.5px !important;
+}
+
+/* Selected state - same color, just thicker */
+.ridgeline-row-selected .ridgeline-line {
+    stroke-width: 3px !important;
+}
+
+/* Axis styling */
+.ridgeline-y-axis text {
+    fill: var(--ridgeline-axis-color);
+    font-size: 11px;
+    font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, sans-serif;
+}
+
+.ridgeline-y-axis line,
+.ridgeline-y-axis path {
+    stroke: var(--ridgeline-axis-color);
+}
+
+/* X-axis styling */
+.ridgeline-x-axis text {
+    fill: var(--ridgeline-axis-color);
+    font-size: 11px;
+    font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, sans-serif;
+}
+
+.ridgeline-x-axis line,
+.ridgeline-x-axis path {
+    stroke: var(--ridgeline-axis-color);
+}
+
+/* Y-axis label hover highlight - bold, no color change */
+.ridgeline-y-label-hover {
+    font-weight: bold;
+}
+
+/* Y-axis label selected state - bold, no color change */
+.ridgeline-y-label-selected {
+    font-weight: bold;
+}
+
+/* Axis labels */
+.ridgeline-axis-label {
+    fill: var(--ridgeline-label-color);
+    font-size: 12px;
+    font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, sans-serif;
+}
