+++ zensical.toml
@@ -34,11 +34,10 @@ nav = [
         { ImageRefreshWidget = "reference/image-refresh.md" },
         { HTMLRefreshWidget = "reference/html-refresh.md" },
         { ProgressBar = "reference/progress-bar.md" },
-        { PulsarChart = "reference/pulsar-chart.md" },
+        { RidgelineChart = "reference/ridgeline-chart.md" },
         { ModuleTreeWidget = "reference/module-tree.md" },
         { Neo4jWidget = "reference/neo4j-widget.md" },
         { ScatterWidget = "reference/scatter-widget.md" },
-        { DiffViewer = "reference/diff-viewer.md" },
         { SplineDraw = "reference/spline-draw.md" },
         { TextCompare = "reference/text-compare.md" },
         { EnvConfig = "reference/env-config.md" },
